requirements.txt
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
backoff==2.2.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
Django==3.2
django-celery-beat==2.5.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.14.0
gql==3.4.0
graphql-core==3.2.3
idna==3.4
kombu==4.6.11
multidict==6.0.2
prompt-toolkit==3.0.38
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
redis==4.5.4
requests==2.28.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.12
vine==1.3.0
wcwidth==0.2.6
yarl==1.8.1
